requirements.txt
BeautifulSoup==3.2.1
Django==1.8.4
argparse==1.2.1
psycopg2==2.6.1
requests==2.8.1
wsgiref==0.1.2
Pillow==3.0.0
sorl-thumbnail==12.3
django-filter==0.11.0
Markdown==2.6.2
djangorestframework==3.2.3

